pyproject.toml
[project]
name = "project-tcc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "google-adk>=1.16.0",
    "icecream>=2.1.8",
    "ipykernel>=7.0.1",
    "litellm>=1.78.3",
    "markdown>=3.6",
    "playwright>=1.58.0",
    "python-dotenv>=1.0.1",
    "pylint>=4.0.1",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.2",
    "rich>=14.2.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.37.0",
    "yfinance>=0.2.66",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
addopts = "-m 'not integration'"
testpaths = ["tests"]
markers = [
    "integration: marks tests that call external LLM APIs (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
branch = true
source = [
    "db",
    "models",
    "orchestrator_agent",
    "repositories",
    "routers",
    "services",
    "tools",
    "utils",
]
omit = [
    "tests/*",
    "scripts/*",
    "notebooks/*",
    ".venv/*",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.coverage.html]
directory = "artifacts/coverage-html"

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["."]
force-exclude = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "ANN",
    "ARG",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "D",
    "DTZ",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FA",
    "FIX",
    "G",
    "I",
    "ICN",
    "ISC",
    "LOG",
    "N",
    "PERF",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "RET",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TC",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "COM812",
    "ISC001",
]
fixable = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
"**/test_*.py" = ["S101"]

[tool.bandit]
exclude_dirs = [".venv", "notebooks"]

[tool.bandit.assert_used]
skips = ["*_test.py", "*test_*.py"]

[tool.ty.environment]
python-version = "3.12"
python-platform = "all"
root = ["."]

[tool.ty.analysis]
allowed-unresolved-imports = []
replace-imports-with-any = []
respect-type-ignore-comments = false

[tool.ty.rules]
all = "error"
deprecated = "error"
unused-ignore-comment = "error"
unused-type-ignore-comment = "error"

[tool.ty.src]
respect-ignore-files = true

[tool.ty.terminal]
error-on-warning = true

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "pytest-cov>=7.0.0",
    "ty>=0.0.26",
]
